requirements.txt
python-dotenv
redis
aiodns==1.1.1
aiohttp==3.6.2
aniso8601==8.0.0
async-timeout==3.0.1
attrs==19.3.0
ccxt==1.18.1360
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-RESTful==0.3.7
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
multidict==4.5.2
pycares==3.0.0
pycparser==2.19
pytz==2019.3
requests==2.22.0
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
yarl==1.1.0
